+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Change Log
 _2023-03-27_
 
 * Replace Koin with Kotlin Inject.
+* Refactor notifications.
 
 ## Version 1.3.9